contracts: tested and fixed VC contracts

+++ contracts/contracts/Reputation.sol
@@ -6,13 +6,11 @@ import "./interfaces/IReputation.sol";
 contract Reputation is IReputation {
     mapping(address => ReputationPoint) public reputation;
 
-    function setReputation(address _owner, uint256 _reputation) internal {
+    function setReputation(address _owner, uint _reputation) internal {
         reputation[_owner].totalReputationPoint = _reputation;
     }
 
-    function getReputationByOwner(
-        address _owner
-    ) public view returns (uint256) {
+    function getReputationByOwner(address _owner) public view returns (uint) {
         return reputation[_owner].totalReputationPoint;
     }
 
@@ -25,15 +23,15 @@ contract Reputation is IReputation {
         }
         reputation[_owner].socialAccountPointStatus[_socialPointType] = true;
         reputation[_owner].socialAccountPoint += 10;
-        uint256 currentReputationPoint = reputation[_owner]
-            .totalReputationPoint + 10;
+        uint currentReputationPoint = reputation[_owner].totalReputationPoint +
+            10;
         setReputation(_owner, currentReputationPoint);
 
         emit ReputationPointUpdated(_owner, currentReputationPoint);
     }
 
     function addVCPoint(address _owner, bytes32 vcHash) external override {
-        uint256 currentReputationPoint;
+        uint currentReputationPoint;
 
         if (reputation[_owner].verifiableCredentials[msg.sender].length > 0) {
             revert VCAlreadyAdded();
@@ -49,8 +47,8 @@ contract Reputation is IReputation {
 
     function addHistoricalTxPoint(address _owner) external override {
         reputation[_owner].transactionHistoryPoint += 1;
-        uint256 currentReputationPoint = reputation[_owner]
-            .totalReputationPoint + 1;
+        uint currentReputationPoint = reputation[_owner].totalReputationPoint +
+            1;
         setReputation(_owner, currentReputationPoint);
 
         emit ReputationPointUpdated(_owner, currentReputationPoint);
